Fixes for casting problems detected on a SuSE 7.2 Itanium, mostly gcc -Wall complaining about the argument of %p not being a void *.
p4raw-id: //depot/perl@13659

+++ regcomp.c
@@ -3187,7 +3187,7 @@ tryagain:
 	 STRLEN newlen = SvCUR(sv);
 	 if (!SIZE_ONLY) {
 	      DEBUG_r(PerlIO_printf(Perl_debug_log, "recode %*s to %*s\n",
-				    oldlen, STRING(ret), newlen, s));
+				    (int)oldlen, STRING(ret), (int)newlen, s));
 	      Copy(s, STRING(ret), newlen, char);
 	      STR_LEN(ret) += newlen - oldlen;
 	      RExC_emit += STR_SZ(newlen) - STR_SZ(oldlen);
